Update News & Events Callout to push content down
