+++ components/EventCallout/EventsCallout.styles.js
@@ -1,6 +1,5 @@
-import styled from 'styled-components';
 import { themeGet } from '@styled-system/theme-get';
-
+import styled from 'styled-components';
 import { system } from 'ui-kit';
 
 const EventsCallout = styled.div`
@@ -8,8 +7,6 @@ const EventsCallout = styled.div`
   border-radius: 0;
   box-shadow: 0px 20px 48px rgba(0, 0, 0, 0.04);
   padding: ${themeGet('space.m')} ${themeGet('space.l')};
-  right: 0;
-  ${props => (props.floater ? 'bottom: 0' : 'top: 0')};
 
   @media screen and (min-width: ${themeGet('breakpoints.md')}) {
     padding: ${themeGet('space.l')} ${themeGet('space.xxl')};
@@ -19,8 +16,6 @@ const EventsCallout = styled.div`
     border-radius: ${themeGet('radii.image')};
     margin-top: 0px;
     padding: ${themeGet('space.m')};
-    right: 50%;
-    transform: translate(100%, 0);
   }
 
   ${system}
