+++ pages/connect/[page].js
@@ -1,19 +1,21 @@
-import { useRouter } from 'next/router';
-
-import { GET_CONTENT_CHANNEL } from 'hooks/useContentChannel';
 import {
   ArticleLink,
   ArticleLinks,
   CampusFilter,
-  EventCallout,
-  EventsCallout,
   Layout,
   MainPhotoHeader,
   MarketingHeadline,
-  MeetTheStaff,
   PageSplit,
   Quote,
 } from 'components';
+import IDS from 'config/ids';
+import { GET_CAMPUSES } from 'hooks/useCampuses';
+import { GET_CONTENT_CHANNEL } from 'hooks/useContentChannel';
+// import { GET_STAFF } from 'hooks/useStaff';
+import { GET_MINISTRY_CONTENT } from 'hooks/useMinistryContent';
+import { GET_UNIVERSAL_CONTENT_ITEM_BY_SLUG } from 'hooks/useUniversalContentItemBySlug';
+import { initializeApollo } from 'lib/apolloClient';
+import { useRouter } from 'next/router';
 import { Button, CardGrid, Longform, Section } from 'ui-kit';
 import {
   getChildrenByType,
@@ -21,14 +23,6 @@ import {
   getMetaData,
   getSlugFromURL,
 } from 'utils';
-import IDS from 'config/ids';
-import { initializeApollo } from 'lib/apolloClient';
-import { Info } from 'phosphor-react';
-import { useTheme } from 'styled-components';
-// import { GET_STAFF } from 'hooks/useStaff';
-import { GET_MINISTRY_CONTENT } from 'hooks/useMinistryContent';
-import { GET_CAMPUSES } from 'hooks/useCampuses';
-import { GET_UNIVERSAL_CONTENT_ITEM_BY_SLUG } from 'hooks/useUniversalContentItemBySlug';
 
 export default function Page({
   data = {},
@@ -38,7 +32,6 @@ export default function Page({
   dropdownData,
 }) {
   const router = useRouter();
-  const theme = useTheme();
 
   const { loading, error, getContentBySlug: node } = data;
 
@@ -56,8 +49,8 @@ export default function Page({
   );
 
   const story = stories.length ? stories[0] : null;
-  const ctaLinks = node.ctaLinks.filter((cta) => cta.image?.sources?.[0]?.uri);
-  const extraCTA = node.ctaLinks.filter((cta) => !cta.image?.sources?.[0]?.uri);
+  const ctaLinks = node.ctaLinks.filter(cta => cta.image?.sources?.[0]?.uri);
+  const extraCTA = node.ctaLinks.filter(cta => !cta.image?.sources?.[0]?.uri);
 
   let links = relatedContent?.getMinistryContent?.length
     ? relatedContent.getMinistryContent.slice(0, 4)
@@ -87,39 +80,8 @@ export default function Page({
               : 0,
           lg: 'xxl',
         }}
+        events={links?.length ? links : null}
       />
-      {links?.length ? (
-        <Section contentProps={{ p: '0 !important' }}>
-          <EventsCallout
-            mx={{ _: 0, lg: 'xl' }}
-            mb={{ _: 'l', md: 'xxl' }}
-            my={{ lg: `-${theme.space.xxl}` }}
-            title="News & Events"
-            icon={
-              <Info
-                size={24}
-                style={{
-                  color: theme.colors.neutrals[900],
-                  opacity: '60%',
-                  marginRight: theme.space.xxs,
-                }}
-              />
-            }
-          >
-            {links.map(link => (
-              <EventCallout
-                key={link.id}
-                title={link.title}
-                description={link.subtitle}
-                imageSrc={link.coverImage?.sources?.[0]?.uri}
-                onClick={() =>
-                  router.push(`/${getSlugFromURL(link?.sharing?.url)}`)
-                }
-              />
-            ))}
-          </EventsCallout>
-        </Section>
-      ) : null}
       {node.htmlContent && (
         <Section>
           <Longform
